requirements.txt
pdflatex>=0.1.3
pocketbase-async>=0.9.0
python-dotenv>=1.0.1